DestroysAssignmentWeights action implemented
