+++ app/services/cancels_course_membership/destroys_assignment_weights.rb
@@ -1,0 +1,17 @@
+module Services
+  module Actions
+    class DestroysAssignmentWeights
+      extend LightService::Action
+
+      expects :membership
+
+      executed do |context|
+        membership = context[:membership]
+
+        AssignmentWeight.for_course(membership.course)
+          .for_student(membership.user)
+          .destroy_all
+      end
+    end
+  end
+end
